pyproject.toml
[tool.poetry]
name = "mwra-data"
version = "0.1.0"
description = "Format-shift MWRA Biobot COVID-19 wastewater data"
authors = ["Ben Steinberg <ben@potswift.org>"]
readme = "README.md"
packages = [{include = "mwra_data"}]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.10"
camelot-py = "^0.11.0"
opencv-python = "^4.8.0.76"
ghostscript = "^0.7"
click = "^8.1.7"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"

[tool.poetry.scripts]
extract = 'mwra_data.extract:get_data'

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
